Base Classes Created + Tree Data Structures Added For Base Classes(Not Including Character)
+ Game.cs, Character.cs, Mage.cs, Rogues.cs, Tank.cs,
TreeDataStructure.cs, TextReadInForTrees.cs created

+ MageTree.txt, RogueTree.txt, TankTree.txt created

————————

+Character.cs includes the stats that will determine everything about
our Mage, Tank and Rogue classes. Accessors and Mutators added.

+ Mage, Rogue and Tank classes contain the number of classes as well as
the TreeDataStructure we will use. addTo[Character]Tree adds the index
of where they are at in the tree as well as the data to be stored.
printTree simply prints the tree in order.

+ Game includes static class variables of the Tank, Mage and Rogue.

+ TextReadInForTrees simply reads in the three text files included in
this commit to create the nodes for the trees.

+ TreeDataStructure is the class that deals with the generation of the
data structure.